2.2.2

+++ share/fs-uae/shaders/scanlines-nonlinear.shader
@@ -16,15 +16,15 @@ uniform vec2 rubyTextureSize;
 
 void main(void) {   
     vec4 rgb = texture2D(rubyTexture, gl_TexCoord[0].xy);
-	int line = int(gl_FragCoord.y);
-	if (int(mod(line, 2)) == 0) {
-		gl_FragColor = mix(rgb, vec4(1.0, 1.0, 1.0, 1.0), 0.05);
-	}
-	else {
+    float line = float(int(gl_FragCoord.y));
+    if (int(mod(line, 2.0)) == 0) {
+        gl_FragColor = mix(rgb, vec4(1.0, 1.0, 1.0, 1.0), 0.05);
+    }
+    else {
         float ia = (128.0 - 10.0) / 255.0 +
-        		((rgb.r + rgb.g + rgb.b) / 3.0) / 2.0;
-		gl_FragColor = rgb * vec4(ia, ia, ia, 1.0);
-	}
+                ((rgb.r + rgb.g + rgb.b) / 3.0) / 2.0;
+        gl_FragColor = rgb * vec4(ia, ia, ia, 1.0);
+    }
 }
 
 ]]></fragment>
